[add] full api

+++ src/main/java/vn/xteam/savemoneyapi/entities/v1/SavingBookEntity.java
@@ -39,7 +39,7 @@ public class SavingBookEntity implements Serializable {
 
     @SerializedName("type")
     @JsonProperty("type")
-    private int type;
+    private Long type;
 
     @SerializedName("created_by")
     @JsonProperty("created_by")
